+++ pentobi/MenuRecentFiles.qml
@@ -29,63 +29,63 @@ PentobiMenu {
     }
 
     // Instantiator in Menu doesn't work reliably with Qt 5.11 or 5.12.0 alpha
-    MenuItem {
+    PentobiMenuItem {
         visible: recentFiles.entries.length > 0
         // Invisible menu item still use space in Qt 5.11
         height: visible ? implicitHeight : 0
         text: getText(recentFiles.entries, 0)
         onTriggered: openFile(0)
     }
-    MenuItem {
+    PentobiMenuItem {
         visible: recentFiles.entries.length > 1
         height: visible ? implicitHeight : 0
         text: getText(recentFiles.entries, 1)
         onTriggered: openFile(1)
     }
-    MenuItem {
+    PentobiMenuItem {
         visible: recentFiles.entries.length > 2
         height: visible ? implicitHeight : 0
         text: getText(recentFiles.entries, 2)
         onTriggered: openFile(2)
     }
-    MenuItem {
+    PentobiMenuItem {
         visible: recentFiles.entries.length > 3
         height: visible ? implicitHeight : 0
         text: getText(recentFiles.entries, 3)
         onTriggered: openFile(3)
     }
-    MenuItem {
+    PentobiMenuItem {
         visible: recentFiles.entries.length > 4
         height: visible ? implicitHeight : 0
         text: getText(recentFiles.entries, 4)
         onTriggered: openFile(4)
     }
-    MenuItem {
+    PentobiMenuItem {
         visible: recentFiles.entries.length > 5
         height: visible ? implicitHeight : 0
         text: getText(recentFiles.entries, 5)
         onTriggered: openFile(5)
     }
-    MenuItem {
+    PentobiMenuItem {
         visible: recentFiles.entries.length > 6
         height: visible ? implicitHeight : 0
         text: getText(recentFiles.entries, 6)
         onTriggered: openFile(6)
     }
-    MenuItem {
+    PentobiMenuItem {
         visible: recentFiles.entries.length > 7
         height: visible ? implicitHeight : 0
         text: getText(recentFiles.entries, 7)
         onTriggered: openFile(7)
     }
-    MenuItem {
+    PentobiMenuItem {
         visible: recentFiles.entries.length > 8
         height: visible ? implicitHeight : 0
         text: getText(recentFiles.entries, 8)
         onTriggered: openFile(8)
     }
     MenuSeparator { }
-    Action {
+    PentobiMenuItem {
         //: Menu item for clearing the recent files list
         text: qsTr("Clear List")
         // recentFiles.clear() must be called after menu is closed because it