+++ pentobi/PentobiMenuItem.qml
@@ -1,19 +1,14 @@
 //-----------------------------------------------------------------------------
-/** @file pentobi/MenuHelp.qml
+/** @file pentobi/PentobiMenuItem.qml
     @author Markus Enzenberger
     @copyright GNU General Public License version 3 or later */
 //-----------------------------------------------------------------------------
 
+import QtQuick
 import QtQuick.Controls
 
-PentobiMenu {
-    title: qsTr("Help")
-
-    PentobiMenuItem {
-        action: actionHelp
-    }
-    PentobiMenuItem {
-        text: qsTr("About Pentobi")
-        onTriggered: aboutDialog.open()
-    }
+MenuItem {
+    // Add extra transparency as a workaround for QTBUG-119198 (Disabled menu
+    // item indistinguishable from enabled one).
+    opacity: workaroundQtBug119198 && ! enabled ? 0.65 : 1
 }